implement update profile

+++ src/core/use-cases/vehicle/remove.ts
@@ -1,19 +1,23 @@
 import { MissingParamError, ServerError } from "../../../utils/errors"
 
 export default function makeRemove({
-    vehicleDb
+    vehicleDb,
+    saveProfile
 }: any = {}) {
-    if (!vehicleDb) throw new ServerError()
+    if (!vehicleDb || !saveProfile) throw new ServerError()
     return async ({
-        id
+        id,
+        userId
     }: any = {}) => {
         if (!id) throw new MissingParamError('id')
+        if (!userId) throw new MissingParamError('userId')
 
         await vehicleDb.deleteOne({ 
             where: {
                 id
             }
         })
+        saveProfile(userId)
         const message = { text: 'response.remove' }
         return { message }
     }
